Refactor to pull out common test functionality
