+++ e2e/init.js
@@ -6,11 +6,11 @@ before(async () => {
   await detox.init(config);
 });
 
-beforeEach(async function () {
+beforeEach(async function() {
   await adapter.beforeEach(this);
 });
 
-afterEach(async function () {
+afterEach(async function() {
   await adapter.afterEach(this);
 });
